% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{addSwitch}
\alias{addSwitch}
\title{Adding recombination to the haplotype matrix}
\usage{
addSwitch(haplotypeMatrix, switchPoints, minLength)
}
\arguments{
\item{haplotypeMatrix}{A matrix with haplotype information where rows are number of animals and columns are number of markers.}

\item{switchPoints}{A list of recombination points per animal.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}
}
\value{
Haplotype matrix with recombination
}
\description{
Adding recombination to the haplotype matrix
}
\examples{
haplotype <- matrix(c(0, 0, 0, 0,
1, 1, 1, 1,
0, 0, 1, 1,
1, 1, 0, 0,
1, 1, 1, 1,
0, 0, 0, 0), byrow = TRUE, nrow = 6)
switchPoints <- list(firstInd = c(2), secondInd = c(1, 3), latsInd = 0)
addSwitch(haplotype, switchPoints, 0)
}
