% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{Admixplot}
\alias{Admixplot}
\title{Plot for True Breed Composition}
\usage{
Admixplot(trueBC, color, legendlabels)
}
\arguments{
\item{trueBC}{True breed composition estimated by the haplotype information coded
according to the breed of origin.}

\item{color}{Colors specifying the number of populations contributing to the true breed composition.}

\item{legendlabels}{Population label to add as a legend to the plot.}
}
\value{
Returns an admixture plot of animals where x axis shows number of animals
and y axis shows the proportion of different breeds in different colors.
}
\description{
Generates barplot of true breed composition of the animals with haplotype information
coded according to the breed of origin.
}
\examples{
popAdm <- matrix(c(-1, 2, 3, -4, 2,
1,-1,2,2,3,
1,2,-3,-4,-5,
2,-3,2,4,-5), byrow = TRUE, nrow = 4, ncol = 5)
TBC <- trueBreedComp(Haplotype = popAdm)
Admixplot(trueBC = TBC, color = c("#FFC107", "#2E7D32", "#FB8072", "#386CB0", "#4d4a49"),
legendlabels = c("A", "B", "C", "D", "E"))
}
\seealso{
\code{\link[=trueBreedComp]{trueBreedComp()}}
}
