% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{AlleleFreq}
\alias{AlleleFreq}
\title{Calculate Allele Frequency}
\usage{
AlleleFreq(df)
}
\arguments{
\item{df}{A matrix consisting of a genotype information of the animals in (0,1,2) format.}
}
\value{
Returns the value of allele frequency at each locus.
}
\description{
Function to calculate allele frequency at each locus of a population.
}
\examples{
popC <- matrix(data = sample(c(0,1,2), 60, replace = TRUE), nrow = 6, ncol = 10)
allelefreq <- AlleleFreq(df = popC)
}
