% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breed_Effect_Function.R
\name{BreedEff_Haplo}
\alias{BreedEff_Haplo}
\title{True Breeding Value (TBV) estimation according to breed of origin}
\usage{
BreedEff_Haplo(Haplotype, Breed_Effect)
}
\arguments{
\item{Haplotype}{A matrix comprising a coded haplotype information of the population where rows are number of animals
and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{Breed_Effect}{A matrix comprising the breed effect defined according to the breed of origin
such that row depicts the breeds and columns depicts the SNP effects.}
}
\value{
Returns a matrix of coded Haplotype with TBV estimated according to the breed of origin.
}
\description{
Function to estimate TBV for animals having different SNP effects according to breed of origin.
}
\examples{
pop_BH <- matrix(sample(c(-1,1,-2,2,-3,3), 16, replace = TRUE), nrow = 4, byrow = TRUE)
Breed_Eff <- matrix(runif(4 * 4, 0, 1), nrow = 4, byrow = TRUE)
Breed_Hap <- BreedEff_Haplo(Haplotype = pop_BH, Breed_Effect = Breed_Eff)

}
