% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breed_Effect_Function.R
\name{BreedPopSelect}
\alias{BreedPopSelect}
\title{Selecting Population with breed effect according to breed of origin}
\usage{
BreedPopSelect(
  map,
  Haplotype,
  nSire,
  nDam,
  rate = 2.5 * 10^-5,
  nProgeny,
  SelType = "random",
  Breed_Effect,
  h2,
  trait_mean,
  VarE,
  recL,
  nChr,
  recProb,
  minLength
)
}
\arguments{
\item{map}{Map file for the population comprising of marker ID, chromosome number and marker position.}

\item{Haplotype}{A matrix with coded haplotype information of the population where rows are number of animals
and columns are number of markers.
In case of haplotype, two rows provide information on genomic material of an animal.}

\item{nSire}{Number of sires to be used for simulation.}

\item{nDam}{Number of dams to be used for simulation.}

\item{rate}{Mutation rate to introduce mutation in the simulated population.}

\item{nProgeny}{Number of progeny to be simulated in each generation.}

\item{SelType}{A character value specifying the selection criteria of the simulated population.
Selection criteria can be based on phenotype, random or true breeding value. The default selection type is "random".
For phenotypic selection, use "Pheno" and for true breeding value, use "TBV".}

\item{Breed_Effect}{A matrix comprising breed effect according to the breed of origin where rows contain information
on SNP effects.}

\item{h2}{A numeric value describing the heritability of the trait to be simulated ranging from 0 to 1.}

\item{trait_mean}{A positive integer value describing mean of the phenotype of the trait to be simulated.}

\item{VarE}{A positive integer value depicting error variance obtained after rescaling the QTL effects.}

\item{recL}{A numeric value describing the average number of recombinations to be introduced per animal.}

\item{nChr}{A numeric value describing the number of haploid chromosomes (N) defined in the simulated population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}
}
\value{
Returns selected sire and dams along with with their ID, recombination points of sire and dam, and population coded haplotype with phenotype and TBV.
}
\description{
Function to simulate a population with different breed effect according to breed of origin.
}
\examples{
PopS <- matrix(data = sample(c(-1,1, -2,2), 200, replace = TRUE),
nrow = 20, ncol = 10)
PopS_map <- generateMAP(2, 5, 5)
PopS_QTL <-  matrix(c(-0.001, 0.0003, 0, 0, 0, 0.0032, 0.0023, -0.0012,0, 0,
 -0.002, 0.004, 0, 0, 0.002, 0, 0.002, 0,0, -0.0012), byrow = 2, nrow = 2, ncol = 10)
Selpop <- BreedPopSelect(map = PopS_map, Haplotype = PopS, nSire = 2, nDam = 2, rate = 2.5 * 10^-5,
nProgeny = 6, SelType = "random", Breed_Effect = PopS_QTL, trait_mean = 5,VarE = 0.6,
h2 = 0.2, recL = 1, nChr = 2, minLength = 0)

}
\seealso{
\code{\link[=BreedEff_Haplo]{BreedEff_Haplo()}}
}
