% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{Coded_to_Haplo}
\alias{Coded_to_Haplo}
\title{Recode Coded haplotypes according to breed of origin into 0 and 1 format}
\usage{
Coded_to_Haplo(df)
}
\arguments{
\item{df}{A matrix consisting of coded haplotype information of the animals
according to the breed of origin.}
}
\value{
Returns a matrix consisting of a haplotype information in (0,1) format.
}
\description{
Function to recode the coded haplotype back to its original form.
}
\examples{
popB <- matrix(data = sample(c(-1,1), 60, replace = TRUE), nrow = 6, ncol = 10)
coded_haplo <- Coded_to_Haplo(df = popB)
}
