% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{DamSample}
\alias{DamSample}
\title{Select Dams}
\usage{
DamSample(Haplotype, nDam, nProgeny)
}
\arguments{
\item{Haplotype}{Haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{nDam}{Number of Dams to be selected as parents for the simulation.}

\item{nProgeny}{Number of progeny to be simulated in the population.}
}
\value{
Return a list of dam IDs and haplotype information of the dam.
}
\description{
Function to select the dams.
}
\examples{
pop_J <- matrix(data = sample(c(0,1), 60, replace = TRUE), nrow = 6, ncol = 10)
pop_J_dam <- DamSample(Haplotype = pop_J, nDam = 1, nProgeny = 4)
}
