% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Phenotype.R
\name{Ext_Pheno}
\alias{Ext_Pheno}
\title{Extract TBV and Phenotype}
\usage{
Ext_Pheno(df)
}
\arguments{
\item{df}{A matrix comprising the haplotype information along with TBV and Phenotype.
Ensure that rownames of the matrix is Animal ID}
}
\value{
A datafile comprising Animal ID, TBV and phenotype.
}
\description{
Extract TBV and Phenotype
}
\examples{
popEP <- matrix(data= sample(c(0,1), 16, replace= TRUE), nrow = 4, ncol = 4)
popEP <- PopulationID(Population = popEP, PopCode = "EP")
QTLEff <- c(0.2, 0.1, 0.3, 0.1)
TBV_haplotype2 <- TBV_Haplo(Haplotype = popEP, Effect = QTLEff)
Var_Ph_R <- Var_PhenoRecords(Haplotype = TBV_haplotype2, h2 = 0.3,
trait_mean = 20, VarE = 0.6)
Haplo_TBV_Ph <- Var_Ph_R$Haplo_pheno
Datafile <- Ext_Pheno(Haplo_TBV_Ph)

}
