% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{generateHP}
\alias{generateHP}
\title{Generate Historical Population}
\usage{
generateHP(
  n_generations,
  n_animals,
  map,
  nSire,
  nDam,
  nProgeny,
  mutationRate = 2.5 * 10^-5,
  SelType = "random",
  Effect,
  h2 = 0.3,
  trait_mean = 10,
  VarE,
  recL = 5,
  nChr,
  recProb,
  minLength = 0,
  IndStartVal = 1,
  prefixID = "H"
)
}
\arguments{
\item{n_generations}{Number of generations to be simulated.}

\item{n_animals}{Number of animals to be simulated.}

\item{map}{Map file for the population.}

\item{nSire}{Number of sires to be selected as parents for the simulation.}

\item{nDam}{Number of Dams to be selected as parents for the simulation.}

\item{nProgeny}{Number of progeny to be simulated in the population.}

\item{mutationRate}{Mutation rate to added to the simulated population.}

\item{SelType}{Selection Type. It can be based on phenotype, random or true breeding value. The default selection type is random.
For phenotypic selection, use "Pheno" and for true breeding value, use "TBV".}

\item{Effect}{SNP effect for calculation of TBV.}

\item{h2}{Heritability of the trait to be simulated.}

\item{trait_mean}{Mean of the trait to be simulated.}

\item{VarE}{Error Variance obtained after rescaling the QTL effects.}

\item{recL}{Average number of recombinations to be introduced per animal.}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}

\item{IndStartVal}{Starting value for the Animal ID of the historical population.}

\item{prefixID}{Prefix to be added to the ID of the animal of the historical population}
}
\value{
Returns haplotype information of the historical population.
}
\description{
Function to simulate historical population.
}
\examples{
HP_map <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 5)
QTL_HP <- QTLeffects(n_QTL_Chr = 2, nChr = 2, len_Chr = 5, shape_gamma = 0.4,
effect_distribution = "gamma")
HP_mapqtl <- mapQTLfile(map = HP_map, QTL = QTL_HP)
HP_effect <- as.numeric(HP_mapqtl$Effect)
H1 <- generateHP(n_generations = 5, n_animals = 20, map = HP_mapqtl,
nSire = 1, nDam = 1, nProgeny = 6, mutationRate = 2.5 * 10^-5, SelType = "random",
Effect = HP_effect, h2 = 0.3, trait_mean = 10, VarE = 0.6,recL = 2,nChr = 2, minLength = 0)

}
\seealso{
\code{\link[=MutationRate]{MutationRate()}}, \code{\link[=mapQTLfile]{mapQTLfile()}}, \code{\link[=generateMAP]{generateMAP()}}, \code{\link[=QTLeffects]{QTLeffects()}}
}
