% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{generateMAP}
\alias{generateMAP}
\title{Generate MAP file}
\usage{
generateMAP(nChr, n_markers, len_Chr)
}
\arguments{
\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{n_markers}{Number of SNP markers to be generated per chromosome.
This parameter is independent of QTL effects at this stage.}

\item{len_Chr}{Length of the chromosome. It can be a single value or vector specifying different length per chromosome in cM.
If users wants to use base pairs(bp) as a measure to determine position, it can be obtained by bp = cM *10^6 .The base pair
position should be calculated by users as it is not part of this function.}
}
\value{
Returns a map file of the population with marker ID, chromosome number and position of the markers on the chromosome.
}
\description{
Function to generate MAP file with marker ID, chromosome and marker position columns.
}
\examples{
map <- generateMAP(nChr = 2, n_markers = 10, len_Chr = 5)
}
\seealso{
\code{\link[=QTLeffects]{QTLeffects()}}, \code{\link[=mapQTLfile]{mapQTLfile()}}
}
