% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePop.R
\name{GenOnePopulation}
\alias{GenOnePopulation}
\title{Simulate Population}
\usage{
GenOnePopulation(
  ngenerations = 5,
  map = map,
  populationSim,
  nSire = 54,
  nDam = 1000,
  mutationRate = 2.5 * 10^-5,
  SelType = "random",
  h2 = 0.3,
  trait_mean,
  VarE,
  Effect,
  recL,
  nChr,
  recProb,
  minLength,
  IndStartVal = 1,
  prefixID = "A",
  nProgeny = 1000
)
}
\arguments{
\item{ngenerations}{Number of generations to be simulated}

\item{map}{Map file for the population comprising of marker ID, chromosome number and marker position.}

\item{populationSim}{Coded haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{nSire}{Number of sires to be selected as parents for the simulation.}

\item{nDam}{Number of Dams to be selected as parents for the simulation.}

\item{mutationRate}{Mutation rate to added to the simulated population.}

\item{SelType}{Selection Type. It can be based on phenotype, random or true breeding value. The default selection type is "random".
For phenotypic selection, use "Pheno" and for true breeding value, use "TBV".}

\item{h2}{Heritability of the trait to be simulated.}

\item{trait_mean}{Mean of the trait to be simulated.}

\item{VarE}{Error Variance obtained after rescaling the QTL effects.}

\item{Effect}{SNP effect for the calculation of TBV.}

\item{recL}{Average number of recombinations to be introduced per animal.}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}

\item{IndStartVal}{Starting value for the Animal ID of the simulated population.}

\item{prefixID}{Prefix to be added to the ID of the animal of simulated population.}

\item{nProgeny}{Number of progeny to be simulated in the population.}
}
\value{
List of Haplotype information , pedigree of the simulated animals, population with TBV and phenotype, recombination point list of sire and dam.
}
\description{
This function simulates a population for multiple number of generations as specified by the users.
It generates coded Haplotype as per breed of origin, from which we can extract haplotype and genotype.It also generates pedigree file,
true breeding value, phenotype, recombination points and recombination probability.
}
\examples{
PopGen <-  matrix(data = sample(c(-1, 1), 200, replace = TRUE), nrow = 20, ncol = 10)
PopGen <- PopulationID(Population = PopGen, PopCode = "A1")
PopGen_map <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 5)
QTL_PopGen <- c(-0.001, 0.0003, 0, 0, 0, 0.0032, 0.0023, -0.0012,0, 0 )
PopF1 <- GenOnePopulation(ngenerations = 2, map = PopGen_map, populationSim = PopGen,
nSire = 1, nDam = 1, mutationRate = 2.5 * 10^-5, SelType = "TBV", h2 = 0.3,
trait_mean = 5,VarE = 0.6, Effect = QTL_PopGen, recL = 1, nChr = 2,
minLength = 0, IndStartVal = 1, prefixID = "A", nProgeny = 10)

}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}, \code{\link[=PopSelect]{PopSelect()}}
}
