% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{MakeGeno}
\alias{MakeGeno}
\title{Make Genotype from Haplotype}
\usage{
MakeGeno(df)
}
\arguments{
\item{df}{Haplotype (0,1) information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}
}
\value{
Returns the genotype information of the animal.
}
\description{
Function to generate genotype information from haplotype (0,1) information.
}
\examples{
popD <- matrix(data = sample(c(0,1), 60, replace = TRUE), nrow = 6, ncol = 10)
genotype <- MakeGeno(df = popD)
}
