% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{mapQTLfile}
\alias{mapQTLfile}
\title{Create combined map and qtl file}
\usage{
mapQTLfile(map, QTL)
}
\arguments{
\item{map}{map file with marker ID, chromosome, and marker position}

\item{QTL}{qtl file with QTL ID, chromosome, marker position and effect}
}
\value{
combined file with marker and QTL ID, chromosome, position and QTL effects.
}
\description{
This function combines marker and qtl file together. SNP markers are often observed to be near or away from the QTL in real life scenario.
However, QTL position is included along with map file to get combined marker file for simulation purpose in this package.
This step is crucial to calculate the true breeding value and simulate the phenotype with desired population parameters.
}
\examples{
popmap <- generateMAP(nChr = 5, n_markers = 10,
len_Chr = c(10, 20, 40, 50, 60))
popQTL <- QTLeffects(n_QTL_Chr = 2, nChr = 5, len_Chr = c(10, 20, 40, 50, 60),
shape_gamma = 0.4, effect_distribution = "gamma")
pop_mapqtl <- mapQTLfile(map = popmap, QTL = popQTL)
}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}, \code{\link[=QTLeffects]{QTLeffects()}}
}
