% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{MatingPop}
\alias{MatingPop}
\title{Function for mating sire and dam}
\usage{
MatingPop(Sire, Dam)
}
\arguments{
\item{Sire}{Haplotype information of the Sire in matrix format.}

\item{Dam}{Haplotype information of the Dam in matrix format.}
}
\value{
Returns haplotype information of the progeny.
}
\description{
Function to generate progeny from selected sires and dams.
}
\examples{
Sire_K <- matrix(data = sample(c(0,1), 30, replace = TRUE), nrow = 3, ncol = 10)
Dam_K <- matrix(data = sample(c(0,1), 30, replace = TRUE), nrow = 3, ncol = 10)
popK <- MatingPop(Sire = Sire_K, Dam = Dam_K)
}
