% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{MutationRate}
\alias{MutationRate}
\title{Adding Mutation}
\usage{
MutationRate(Haplotype, rate)
}
\arguments{
\item{Haplotype}{Haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{rate}{Mutation rate. The default mutation rate is = 2.5*10^-5.
However, user can also define mutation rate as per requirement.}
}
\value{
Returns a haplotype matrix with mutations.
}
\description{
Function to add mutation rate to the haplotype information of the animals.
}
\examples{
popH <- matrix(data = sample(c(0,1), 60, replace = TRUE), nrow = 6, ncol = 10)
popH_Mutation <- MutationRate(Haplotype = popH, rate = 2.5*10^-5)
}
