% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Phenotype.R
\name{Ped_Pheno}
\alias{Ped_Pheno}
\title{Combine Pedigree and Phenotype data}
\usage{
Ped_Pheno(Pedigree, Phenotype)
}
\arguments{
\item{Pedigree}{Pedigree file comprising AnimalID, Sire ID, Dam ID and Sex of the animal.}

\item{Phenotype}{Phenotype file comprising Animal ID, TBV and phenotype of the animal}
}
\value{
A datafile comprising AnimalID, Sire ID, Dam ID, Sex, TBV and phenotype of the animal.
}
\description{
Combine Pedigree and Phenotype data
}
\examples{
PopGen1 <-  matrix(data = sample(c(-1, 1), 200, replace = TRUE), nrow = 20, ncol = 10)
PopGen1 <- PopulationID(Population = PopGen1, PopCode = "PG")
PopGen1_map <-  generateMAP(2, 5, 5)
QTL_PopGen1  <-  c(-0.001, 0.0003, 0, 0, 0, 0.0032, 0.0023, -0.0012,0, 0)
PopPP <- GenOnePopulation(ngenerations=2, map=PopGen1_map,
PopGen1,nSire=1,nDam=1,mutationRate=2.5 * 10^-5,
SelType = "TBV", h2 = 0.3, trait_mean = 5 ,VarE = 0.6, Effect = QTL_PopGen1,
recL = 1, nChr = 2, minLength = 0, IndStartVal=1,prefixID="A",nProgeny=10)
Phenofile <- Ext_Pheno(df = PopPP$population)
Ped_Phenofile <- Ped_Pheno(Pedigree = PopPP$parents, Phenotype = Phenofile)


}
\seealso{
\code{\link[=Ext_Pheno]{Ext_Pheno()}}
}
