% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotHaplo}
\alias{plotHaplo}
\title{Plot for Haplotype segments}
\usage{
plotHaplo(Haplotype, colors, map, nChr, legendlabels)
}
\arguments{
\item{Haplotype}{Haplotype information of the animals to be plotted.}

\item{colors}{Colors to be defined for different populations.}

\item{map}{Mapfile of the population comprising of marker ID, chromosome number and marker position.}

\item{nChr}{Haploid number of chromosomes(n).}

\item{legendlabels}{Population label to add as a legend to the plot.}
}
\value{
Haplotype plot showing blocks from different population for animals.
}
\description{
Function to generate plot of haplotype segments inherited from different breeds.It
utilises the coded haplotype information according to the breed of origin and generates plot
at all chromosomes.
}
\examples{
Haplotype <- matrix(data = sample(c(-1, 1, -2, 2, -3, 3), 100, replace = TRUE),
nrow = 10, ncol = 10)
Haplotype_ID <- PopulationID(Population = Haplotype, PopCode = "H1")
map <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 5)
colors = c("gold", "forestgreen", "steelblue")
plot <- plotHaplo(Haplotype = Haplotype_ID, colors = colors, map = map,
nChr = 2, legendlabels = c("A", "B", "C"))
}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}
}
