% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePop.R
\name{PopSelect}
\alias{PopSelect}
\title{Selecting Population with selection criteria random, phenotype and true breeding value}
\usage{
PopSelect(
  map,
  Haplotype,
  nSire,
  nDam,
  rate = 2.5 * 10^-5,
  nProgeny,
  SelType = "random",
  Effect,
  h2,
  trait_mean,
  VarE,
  recL,
  nChr,
  recProb,
  minLength
)
}
\arguments{
\item{map}{Map file for the population comprising of marker ID, chromosome number and marker position.}

\item{Haplotype}{Coded haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{nSire}{Number of Sires to be used for simulation.}

\item{nDam}{Number of Dams to be used for simulation.}

\item{rate}{Mutation rate to introduce mutation in the population.}

\item{nProgeny}{Number of Progeny to be simulated in each generation.}

\item{SelType}{Selection Type. It can be based on phenotype, random or true breeding value. The default selection type is "random".
For phenotypic selection, use "Pheno" and for true breeding value, use "TBV".}

\item{Effect}{SNP effect for the calculation of TBV.}

\item{h2}{Heritability of the trait to be simulated.}

\item{trait_mean}{Mean of the trait to be simulated.}

\item{VarE}{Error Variance obtained after rescaling the QTL effects.}

\item{recL}{Average number of recombinations to be introduced per animal.}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}
}
\value{
It generates selected sire and dams along with with their ID, recombination points of sire and dam, and population coded haplotype with phenotype and TBV.
}
\description{
This function select sires and dams of the population using random, phenotypic or true breeding value.
}
\examples{
PopS <- matrix(data = sample(c(0,1), 200, replace = TRUE),
nrow = 20, ncol = 10)
PopS_map <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 5)
PopS_QTL <-  c(-0.001, 0.0003, 0, 0, 0, 0.0032, 0.0023, -0.0012,0, 0 )
Selpop <-  PopSelect(map = PopS_map, Haplotype = PopS, nSire = 2, nDam = 2,
rate = 2.5 * 10^-5, nProgeny = 6, SelType = "random", Effect = PopS_QTL,
trait_mean = 5, VarE = 0.6, h2 = 0.2, recL = 1, nChr = 2, minLength = 0)

}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}
}
