% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{PopulationID}
\alias{PopulationID}
\title{Adding Population Code before simulating the population}
\usage{
PopulationID(Population, PopCode)
}
\arguments{
\item{Population}{Haplotype information of the base population to be used for simulating the population.}

\item{PopCode}{Code to be assigned for the base population which will appear in IDs.}
}
\value{
Returns a population with provided PopCode, which will appear as PopulationID.
}
\description{
Function to add Population code to the population which will help in generating parent IDs.
}
\examples{
PopN_Check <- matrix(data = sample(c(0,1), 200, replace = TRUE), nrow = 10, ncol = 20)
PopN_CheckID <- PopulationID(Population = PopN_Check, PopCode = "P1")
}
