% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{QTL_densityplot}
\alias{QTL_densityplot}
\title{Plot for the QTL effect}
\usage{
QTL_densityplot(QTL_eff_file, color = "steelblue")
}
\arguments{
\item{QTL_eff_file}{QTL file with information on QTL effects.}

\item{color}{color of the density plot.}
}
\value{
shows density plot depicting the distribution of the QTL effects.
}
\description{
Generates the density plot for the QTL effects.
}
\examples{
example_QTL <- QTLeffects(n_QTL_Chr = 1, nChr = 5, len_Chr = c(10, 8, 6, 4, 2),
shape_gamma = 0.4)
plot <- QTL_densityplot(QTL_eff_file = example_QTL, color = "forestgreen")
}
\seealso{
\code{\link[=QTLeffects]{QTLeffects()}}
}
