% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{QTLeffects}
\alias{QTLeffects}
\title{Calculation QTL effects}
\usage{
QTLeffects(
  n_QTL_Chr,
  nChr,
  len_Chr,
  shape_gamma = 0.4,
  effect_distribution = "gamma"
)
}
\arguments{
\item{n_QTL_Chr}{Number of QTL to be simulated per chromosome}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{len_Chr}{Length of the chromosome.It can be either provided single constant value or vector specifying varied length of the chromosomes.}

\item{shape_gamma}{Shape of the QTL effects if simulated using gamma distribution.}

\item{effect_distribution}{Define the distribution of QTL effects. It can be either gamma, normal or uniform distribution.}
}
\value{
file comprising of QTL ID, chromosomes, position of the QTL and QTL effects.
}
\description{
Function to estimate QTL effects with output of QTL ID, chromosome, QTL position and QTL effects
}
\examples{
QTL <- QTLeffects(n_QTL_Chr = 1, nChr = 2, len_Chr = 10,
shape_gamma = 0.4, effect_distribution = "gamma")
}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}, \code{\link[=mapQTLfile]{mapQTLfile()}}
}
