% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{Random_gametes}
\alias{Random_gametes}
\title{Random selection of sire and dam gametes}
\usage{
Random_gametes(Gamete_Matrix)
}
\arguments{
\item{Gamete_Matrix}{A matrix comprising haplotype information of sire or dam gametes after recombination}
}
\value{
Returns matrix comprising random selection of gametes
}
\description{
Random selection of sire and dam gametes
}
\examples{
mat = matrix(data = sample(c(-1,1,-2,2), 100*10, replace = TRUE)
, nrow = 100, ncol = 10)
mat = PopulationID(Population = mat, PopCode = "M")
rev_mat = Random_gametes(Gamete_Matrix = mat)

}
