% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMSim_functions.R
\name{ReadQTL}
\alias{ReadQTL}
\title{Read QTL file obtained from QMSim}
\usage{
ReadQTL(filename)
}
\arguments{
\item{filename}{Requires the name of the QTL file with allele effect obtained from QMSim.}
}
\value{
Returns a readable QTL effect
}
\description{
This function reads the effect file generated from QMSim with the file prefix "effect_qtl".
It generates the effects for the defined QTLs in QMSim.
}
\examples{
\dontrun{
QTLfile <- ReadQTL(filename = "effect_qtl_001.txt")
}

}
\seealso{
\code{\link[=CalQTLeffect]{CalQTLeffect()}}
}
