% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{Recode_Haplotype}
\alias{Recode_Haplotype}
\title{Recode Haplotype with different breed codes}
\usage{
Recode_Haplotype(df, popBase)
}
\arguments{
\item{df}{A matrix consisting of haplotype information of the population where
rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{popBase}{Code specified according to the specific breed.}
}
\value{
Returns a matrix consisting of coded haplotype information of the animals
according to the defined popBase.
}
\description{
Function to recode the haplotype according to different breed codes as specified by users.
}
\examples{
popA <- matrix(data = sample(c(0,1), 60, replace = TRUE), nrow = 6, ncol = 10)
PA_Haplo <- Recode_Haplotype(df = popA, popBase = 1)
}
