% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{RecombinationPoint}
\alias{RecombinationPoint}
\title{Add Recombination}
\usage{
RecombinationPoint(map, Haplotype, recL, nChr, recProb)
}
\arguments{
\item{map}{Mapfile of the population comprising of marker ID, chromosome number and marker position.}

\item{Haplotype}{Haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{recL}{Average number of recombinations to be introduced per animal.}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}
}
\value{
Returns a list of recombination points and number of recombinations per animal.
}
\description{
Function to add recombination to the haplotype information of the animals.
}
\examples{
popG <- matrix(data = sample(c(0,1), 40, replace = TRUE), nrow = 4, ncol = 10)
popGmap <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 10)
popGRec <- RecombinationPoint(map = popGmap, Haplotype = popG, recL = 2, nChr = 2)
}
