% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{recProbplot}
\alias{recProbplot}
\title{Plot for recombination points and recombination probability}
\usage{
recProbplot(recList, n_markers, col = "steelblue", map)
}
\arguments{
\item{recList}{List of recombination points with each list containing information
of recombination points of an animal.}

\item{n_markers}{Total number of markers.}

\item{col}{Color of the recombination probability plot.}

\item{map}{Map file of the markers for the population.}
}
\value{
This function returns the plots for recombination points and recombination probability.
}
\description{
Function to generate plot for recombination points and recombination probability.
}
\examples{
map_rec <- generateMAP(nChr = 2, n_markers = 5, len_Chr = 10)
poprec <- matrix(data = sample(c(0,1), 1000, replace = TRUE), nrow = 100, ncol = 10)
RecPoint <- RecombinationPoint(map = map_rec, Haplotype = poprec, recL = 2, nChr = 2)
reclist <- RecPoint$recList
recom_plot <- recProbplot(recList = reclist, n_markers = 10, map = map_rec)
}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}, \code{\link[=RecombinationPoint]{RecombinationPoint()}}
}
