% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rescale.R
\name{Rescale}
\alias{Rescale}
\title{Rescaling SNP Effect}
\usage{
Rescale(Haplotype, Effect, Phenosd, h2)
}
\arguments{
\item{Haplotype}{A matrix where each row represents a haplotype (0,1)
and each column represents a marker.}

\item{Effect}{A vector with SNP effect of a population to be scaled before the simulation of a population}

\item{Phenosd}{A value of phenotypic standard deviation which will help in estimating desirable variance components}

\item{h2}{Heritability of the trait to be simulated for the population.}
}
\value{
This function returns a scaled SNP effect, scaled BV along with
phenotypic, additive and residual variance.
}
\description{
This function allows scaling of the SNP effects based on the scaling factor
derived from genotype of the base population and standard deviation of the phenotype.
This allows users to scale the SNP effects to get desirable variance components.
}
\examples{
popRescale <- matrix(c(0,1,1,0,
1,0,0,0,
1,1,1,1,
1,1,1,0), byrow = TRUE, nrow = 4, ncol = 4)
QTLEff = c(0.1, 0.2, 0, 0)
RescaleEff <- Rescale(Haplotype = popRescale, Effect = QTLEff,
Phenosd = 4, h2 = 0.2)
}
