% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Simulation_function.R
\name{SireSample}
\alias{SireSample}
\title{Select Sires}
\usage{
SireSample(Haplotype, nSire, nProgeny)
}
\arguments{
\item{Haplotype}{Haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{nSire}{Number of sires to be selected as parents for the simulation.}

\item{nProgeny}{Number of progeny to be simulated in the population.}
}
\value{
Returns a list of sire IDs and haplotype information of the sire.
}
\description{
Function to select the sire.
}
\examples{
popH <- matrix(data = sample(c(0,1), 60, replace = TRUE), nrow = 6, ncol = 10)
popH_Sire <- SireSample(Haplotype = popH, nSire = 1, nProgeny = 4)
}
