% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{TBV_Haplo}
\alias{TBV_Haplo}
\title{Calculation of True breeding Value (TBV)}
\usage{
TBV_Haplo(Haplotype, Effect)
}
\arguments{
\item{Haplotype}{Haplotype (0,1) information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{Effect}{A vector comprising the SNP effects of the population.}
}
\value{
Returns a haplotype information with TBV estimated from animal's haplotype information.
The TBV mentioned in "TBV" column provides final TBV (sum of TBV from haplotype of the animal) for an animal.
}
\description{
Function to estimate true breeding value.
}
\examples{
popE <- matrix(data= sample(c(0,1), 16, replace= TRUE), nrow = 4, ncol = 4)
QTLEff <- c(0.2, 0.1, 0.3, 0.1)
TBV_haplotype <- TBV_Haplo(Haplotype = popE, Effect = QTLEff)
}
\seealso{
\code{\link[=Var_PhenoRecords]{Var_PhenoRecords()}}
}
