% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{trueBreedComp}
\alias{trueBreedComp}
\title{Calculate True Breed Proportion with coded haplotype}
\usage{
trueBreedComp(Haplotype)
}
\arguments{
\item{Haplotype}{A matrix with coded haplotype information of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}
}
\value{
Returns true breed proportion of population with coded haplotype.
}
\description{
Function to calculate true breed proportion in an animal with coded haplotype information.
}
\examples{
P1 <- matrix(data = sample(c(-1, 1, 2, -2), 20, replace = TRUE), nrow = 4, ncol = 5)
TBP <- trueBreedComp(Haplotype = P1)
}
\seealso{
\code{\link[=Admixplot]{Admixplot()}}
}
