% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePop.R
\name{TwoPopCross}
\alias{TwoPopCross}
\title{Function for crossing two populations}
\usage{
TwoPopCross(
  map = map,
  populationSim_A,
  populationSim_B,
  nSireA = 54,
  nDamA = 1000,
  nSireB = 26,
  nDamB = 1000,
  Sire_From,
  mutationRate = 2.5 * 10^-5,
  SelType,
  h2 = 0.3,
  trait_mean,
  VarE_PopA,
  VarE_PopB,
  Effect_PopA,
  Effect_PopB,
  IndStartVal = 1,
  prefixID = "Cr",
  nProgeny = 1000,
  recL,
  nChr,
  recProb,
  minLength
)
}
\arguments{
\item{map}{Map file for the population comprising of marker ID, chromosome number and marker position.}

\item{populationSim_A}{Coded haplotype information of the population A where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{populationSim_B}{Coded haplotype information of the population B where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{nSireA}{Number of sires to be selected as parents for the simulation from population A.}

\item{nDamA}{Number of dams to be selected as parents for the simulation from population A.}

\item{nSireB}{Number of sires to be selected as parents for the simulation from population B.}

\item{nDamB}{Number of dams to be selected as parents for the simulation from population B.}

\item{Sire_From}{The population from which sires should being selected for cross population.
If sire should be selected from population A, then use "A" and if sire should be selected from population B,
then use "B".}

\item{mutationRate}{Mutation rate to added to the simulated population.}

\item{SelType}{Selection Type. It can be based on phenotype, random or true breeding value. The default selection type is "random".
For phenotypic selection, use "Pheno" and for true breeding value, use "TBV".}

\item{h2}{Heritability of the trait to be simulated.}

\item{trait_mean}{Mean of the trait to be simulated.}

\item{VarE_PopA}{Error Variance obtained after rescaling the QTL effects for PopA.}

\item{VarE_PopB}{Error Variance obtained after rescaling the QTL effects for PopB.}

\item{Effect_PopA}{SNP effect for the calculation of TBV for population A.}

\item{Effect_PopB}{SNP effect for the calculation of TBV for population B.}

\item{IndStartVal}{Starting value for the animal ID.}

\item{prefixID}{Prefix to be added to the ID of the animal.}

\item{nProgeny}{Number of progeny to be simulated in the population.}

\item{recL}{Average number of recombinations to be introduced per animal.}

\item{nChr}{Number of haploid chromosomes (n) defined in the population.}

\item{recProb}{A vector of recombination probability between the markers. Its length should be total
number of markers minus 1. User can define the recombination probability between markers
helping in simulating a desired population.}

\item{minLength}{Minimum length between recombination points beyond which recombination is allowed. The default value is 0.
If a user define minimum length, it will ensure that difference between two recombination points
less than minimum length won't allow recombination.}
}
\value{
List of Haplotype information , ID of the simulated animals, population with TBV and phenotype, recombination point list of sire and dam.
}
\description{
Function for making a cross between two populations.
}
\examples{
crossmap <- generateMAP(nChr = 5, n_markers = 2, len_Chr = 20)
QTL_cross <- c(-0.001, 0.0003, 0, 0, 0, 0.0032, 0.0023, -0.0012,0, 0 )
P1cross <- matrix(data = sample(c(-1, 1), 40, replace = TRUE), nrow = 4, ncol = 10)
P2cross <- matrix(data = sample(c(-2, 2), 40, replace = TRUE), nrow = 4, ncol = 10)
P1cross <- PopulationID(Population = P1cross, PopCode = "P1")
P2cross <- PopulationID(Population = P2cross, PopCode = "P2")
F2cross <- TwoPopCross(map = crossmap, populationSim_A = P1cross, populationSim_B = P2cross,
nSireA = 1, nDamA = 1, nSireB = 1, nDamB = 1, Sire_From = "A", mutationRate = 2.5 * 10^-5,
SelType = "TBV", h2 = 0.3, Effect_PopA = QTL_cross, Effect_PopB = QTL_cross, IndStartVal = 1,
prefixID ="Cr", nProgeny = 10, recL = 1, nChr = 2, minLength = 0, trait_mean = 5,
VarE_PopA = 0.6, VarE_PopB = 0.6)

}
\seealso{
\code{\link[=generateMAP]{generateMAP()}}, \code{\link[=PopSelect]{PopSelect()}}, \code{\link[=RecombinationPoint]{RecombinationPoint()}}
}
