% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{Var_PhenoRecords}
\alias{Var_PhenoRecords}
\title{Generate variance components and phenotypic records}
\usage{
Var_PhenoRecords(Haplotype, h2, trait_mean, VarE)
}
\arguments{
\item{Haplotype}{Haplotype information (0,1) of the population where rows are number of animals and columns are number of markers in matrix form.
In case of haplotype, 2 rows provide information on genomic material of an animal.}

\item{h2}{Heritability of the trait to be simulated.}

\item{trait_mean}{Mean of the trait to be simulated.}

\item{VarE}{Error Variance obtained after rescaling the QTL effects}
}
\value{
Additive genetic variance, phenotypic variance, error variance and haplotype marker information with phenotypic records.
}
\description{
Function to generate variance components and phenotype records for the simulated population.
}
\examples{
popF <- matrix(data= sample(c(0,1), 16, replace= TRUE), nrow = 4, ncol = 4)
QTLEff <- c(0.2, 0.1, 0.3, 0.1)
TBV_haplotype <- TBV_Haplo(Haplotype = popF, Effect = QTLEff)
Var_Ph_R <- Var_PhenoRecords(Haplotype = TBV_haplotype, h2 = 0.3,
trait_mean = 20, VarE = 0.6)
}
\seealso{
\code{\link[=TBV_Haplo]{TBV_Haplo()}}
}
