% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMSim_functions.R
\name{CalQTLeffect}
\alias{CalQTLeffect}
\title{Calculate QTL effect from file generated from ReadQTL (For QMSim generated files)}
\usage{
CalQTLeffect(filename, nSNP, QTLfile)
}
\arguments{
\item{filename}{Requires the name of the file with markerID, chr, and marker position.}

\item{nSNP}{Total number of SNP markers.}

\item{QTLfile}{QTL effects generated from ReadQTL function.}
}
\value{
Returns a vector of the SNP effects with length equals to the total number of SNP markers.
}
\description{
This function requires the file with information on markerID, chromosome and marker information with prefix "lm_mrk_qtl",
total number of SNP markers ,and QTL effects generated from ReadQTL function. It generates
SNP effects of length of total number of markers along with QTLs.
}
\examples{
\donttest{
popQM <- read_qmsim_geno("p1_mrk_qtl_001.txt")
QTL <- ReadQTL(filename = "effect_qtl_001.txt")
QTLeffect <- CalQTLeffect(filename = "lm_mrk_qtl_001.txt",
nSNP = ncol(popQM),QTLfile = QTL)
}


}
\seealso{
\code{\link[=ReadQTL]{ReadQTL()}}
}
