% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{plinkped}
\alias{plinkped}
\title{Generate PLINK pedigree file}
\usage{
plinkped(Genotype, AnimalID, filename)
}
\arguments{
\item{Genotype}{A matrix containing the genotype information of the animals where rows}

\item{AnimalID}{Animal ID for the animals in the population}

\item{filename}{Naming a ped file}
}
\value{
A ped file for PLINK analysis.
As of now, it only saves animal ID and genotypes.
Saving pedigree in ped file is in our plan for the update.
}
\description{
Function to generate ped file for PLINK analysis.
}
\examples{
\dontrun{
pop_plink <- matrix(data = sample(c(0,1,2), 60, replace = TRUE), nrow = 6, ncol = 10)
pop_plink <- PopulationID(Population = pop_plink, PopCode = "PP")
plinkped(Genotype = pop_plink, AnimalID = rownames(pop_plink), filename = "pop")
}


}
