% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{QMSim_PlinkMAP}
\alias{QMSim_PlinkMAP}
\title{Format PLINK MAP file from QMSim}
\usage{
QMSim_PlinkMAP(filepath = "lm_mrk_qtl_001.txt", outputpath, filename)
}
\arguments{
\item{filepath}{Map file generated from the QMSim.}

\item{outputpath}{Path directory for the output file}

\item{filename}{Name of the map file to be extracted}
}
\value{
Returns a map file that can be used for PLINK analysis.
}
\description{
Function to format map file from QMSim for PLINK analysis.
}
\examples{
\dontrun{
QMSim_PlinkMAP(filepath = "lm_mrk_qtl_001.txt",
outputpath = "ss/out", filename = "pop")
}

}
