% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Data_function.R
\name{Reformat_BLUPF90}
\alias{Reformat_BLUPF90}
\title{Reformat genotype data for BLUPF90 analysis}
\usage{
Reformat_BLUPF90(Genodata, savefile)
}
\arguments{
\item{Genodata}{A matrix comprising genotype information (0,1,2) of the animals where rownames depicts the animal ID.}

\item{savefile}{File name in which the genotype data will be saved.}
}
\value{
Returns a text file comprising of genotype data for its use in BLUPF90.
}
\description{
Function to format genotype file for BLUPF90 analysis.
}
\examples{
\dontrun{
Geno <- matrix(data = sample(c(0,1,2), 100, replace = TRUE), nrow = 10, ncol = 10)
Reformat_BLUPF90(Genodata = Geno, savefile = "genotype")
}

}
